# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Consent\u2013Attribution\u2013Settlement Ledger",
    "description": "Tamper-evident infrastructure that makes artist consent, attribution, and compensation mechanically enforceable at the moment of AI generation.",
    "version": "0.1.0"
  },
  "paths": {
    "/rights-holders": {
      "get": {
        "tags": [
          "rights-holders"
        ],
        "summary": "List All",
        "operationId": "list_all_rights_holders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RightsHolderOut"
                  },
                  "type": "array",
                  "title": "Response List All Rights Holders Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rights-holders"
        ],
        "summary": "Create",
        "operationId": "create_rights_holders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RightsHolderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RightsHolderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/artists": {
      "get": {
        "tags": [
          "artists"
        ],
        "summary": "List All",
        "operationId": "list_all_artists_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtistOut"
                  },
                  "type": "array",
                  "title": "Response List All Artists Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "artists"
        ],
        "summary": "Create",
        "operationId": "create_artists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtistCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/generations": {
      "post": {
        "tags": [
          "provenance"
        ],
        "summary": "Post Generation",
        "operationId": "post_generation_generations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/consumption-events": {
      "post": {
        "tags": [
          "consumption"
        ],
        "summary": "Post Event",
        "operationId": "post_event_consumption_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumptionEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumptionEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtistCreate": {
        "properties": {
          "rights_holder_id": {
            "type": "integer",
            "title": "Rights Holder Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "rights_holder_id",
          "name"
        ],
        "title": "ArtistCreate"
      },
      "ArtistOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "rights_holder_id": {
            "type": "integer",
            "title": "Rights Holder Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rights_holder_id",
          "name",
          "created_at"
        ],
        "title": "ArtistOut"
      },
...
      "ConsumptionEventCreate": {
        "properties": {
          "generation_index": {
            "type": "integer",
            "title": "Generation Index"
          },
          "units": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Units"
          },
          "period": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Period"
          }
        },
        "type": "object",
        "required": [
          "generation_index",
          "units",
          "period"
        ],
        "title": "ConsumptionEventCreate"
      },
      "ConsumptionEventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "generation_index": {
            "type": "integer",
            "title": "Generation Index"
          },
          "units": {
            "type": "integer",
            "title": "Units"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "generation_index",
          "units",
          "period",
          "created_at"
        ],
        "title": "ConsumptionEventOut"
      },
...
      "GenerationCreate": {
        "properties": {
          "initiating_artist_id": {
            "type": "integer",
            "title": "Initiating Artist Id"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/ProposedInput"
            },
            "type": "array",
            "minItems": 1,
            "title": "Inputs"
          }
        },
        "type": "object",
        "required": [
          "initiating_artist_id",
          "inputs"
        ],
        "title": "GenerationCreate"
      },
      "GenerationOut": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "initiating_artist_id": {
            "type": "integer",
            "title": "Initiating Artist Id"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/RecordedInput"
            },
            "type": "array",
            "title": "Inputs"
          },
          "prev_hash": {
            "type": "string",
            "title": "Prev Hash"
          },
          "entry_hash": {
            "type": "string",
            "title": "Entry Hash"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          }
        },
        "type": "object",
        "required": [
          "index",
          "timestamp",
          "initiating_artist_id",
          "inputs",
          "prev_hash",
          "entry_hash",
          "signature"
        ],
        "title": "GenerationOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProposedInput": {
        "properties": {
          "artist_id": {
            "type": "integer",
            "title": "Artist Id"
          },
          "use_type": {
            "$ref": "#/components/schemas/UseType"
          },
          "requested_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Requested Weight"
          }
        },
        "type": "object",
        "required": [
          "artist_id",
          "use_type",
          "requested_weight"
        ],
        "title": "ProposedInput"
      },
...
      "RecordedInput": {
        "properties": {
          "artist_id": {
            "type": "integer",
            "title": "Artist Id"
          },
          "rights_holder_id": {
            "type": "integer",
            "title": "Rights Holder Id"
          },
          "use_type": {
            "$ref": "#/components/schemas/UseType"
          },
          "consent_policy_version": {
            "type": "integer",
            "title": "Consent Policy Version"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          }
        },
        "type": "object",
        "required": [
          "artist_id",
          "rights_holder_id",
          "use_type",
          "consent_policy_version",
          "weight"
        ],
        "title": "RecordedInput"
      },
...
      "RightsHolderCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RightsHolderCreate"
      },
      "RightsHolderOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "RightsHolderOut"
      },
...
      "UseType": {
        "type": "string",
        "enum": [
          "STYLE_CONDITIONING",
          "VOICE_CLONING",
          "STEM_USE",
          "FULL_TRACK_REFERENCE",
          "LYRICAL_REFERENCE"
        ],
        "title": "UseType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}